requirements.txt
Flask
Flask-JWT
Flask-RESTful
Flask-SQLAlchemy
mysqlclient
flasgger